requirements.txt
# Core dependencies
anthropic>=0.18.0
openai>=1.12.0
ollama>=0.1.0

# Data validation
pydantic>=2.0.0

# CLI and display
rich>=13.0.0
pyyaml>=6.0.0

# Testing
pytest>=8.0.0
pytest-mock>=3.12.0
